Implemented recycler view

+++ app/src/main/java/me/isachenko/loansonline/data/model/responses/LoanConditionsResponse.kt
@@ -2,6 +2,6 @@ package me.isachenko.loansonline.data.model.responses
 
 data class LoanConditionsResponse(
     val maxAmount: Int,
-    val percent: Int,
+    val percent: Double,
     val period: Int
 )
